+++ R/utils.R
@@ -374,5 +374,9 @@ utils::globalVariables(c(".", ":=", "AnzahlFall", "Area type", "Specimen date", 
                          "cumTestsByPublishDate", "data_list", "newAdmissions", "newCasesByPublishDate",
                          "newCasesBySpecimenDate", "newDeaths28DaysByPublishDate",
                          "newTestsByPublishDate", "setTxtProgressBar", "txtProgressBar",
-                         "level_2_region_code.x", "level_2_region_code.y", "cumDeaths28DaysByDeathDate", "newDeaths28DaysByDeathDate"))
-
+                         "level_2_region_code.x", "level_2_region_code.y", "cumDeaths28DaysByDeathDate", "newDeaths28DaysByDeathDate",
+                         "outcome", "death", "delay_onset_report", "delay_onset_admission", "delay_onset_death",
+                         "hosp_new_first_admissions",
+                         "iso_3166_2", "min_date", "max_date", "iso_na", 
+                         "cases_new_na", "cases_total_na", "deaths_new_na", "deaths_total_na",
+                         "ENTRY_DATE", "DATE_IMPLEMENTED"))